+++ packages/verify/src/index.ts
@@ -20,3 +20,5 @@ export { KeyFinderFunc, TrustMaterial, toTrustMaterial } from './trust';
 export { Verifier, VerifierOptions } from './verifier';
 
 export type { SignedEntity, Signer, VerificationPolicy } from './shared.types';
+
+export type { ObjectIdentifierValuePair } from '@sigstore/protobuf-specs';